+++ src/main/java/com/jug/config/MigrationCostCalculationMethod.java
@@ -1,0 +1,6 @@
+package com.jug.config;
+
+public enum MigrationCostCalculationMethod {
+    ABSOLUTE_POSITION,
+    TOTAL_COMPONENT_LENGTH_BELOW
+}